requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.7
aiosignal==1.3.1
alembic==1.14.0
attrs==24.2.0
cffi==1.17.1
cryptography==44.0.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
jeepney==0.8.0
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
PyNaCl==1.5.0
SecretStorage==3.3.3
SQLAlchemy==2.0.36
typing_extensions==4.12.2
yarl==1.18.0
yt-dlp==2024.11.18
python-dotenv==0.21.0